# lang/codes/gbnzt/docs/swagger.yaml
...
info:
  contact: {}
paths:
...
  /user/api/users:
    delete:
      consumes:
      - application/json
      description: 删除用户
      parameters:
      - in: query
        name: name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: '{"data": nil, "code":200, "message": "删除用户成功"}'
          schema:
            type: string
      tags:
      - 系统用户相关
    get:
      consumes:
      - application/json
      description: 用户查询
      produces:
      - application/json
      responses:
        "200":
          description: '{"data": object, "code":200, "message": "用户查询成功"}'
          schema:
            type: string
      tags:
      - 系统用户相关
    post:
      consumes:
      - application/json
      description: 添加用户
      parameters:
      - in: query
        name: name
        required: true
        type: string
      - in: query
        name: passone
        required: true
        type: string
      - in: query
        name: passtow
        required: true
        type: string
      - in: query
        name: role
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: '{"data": nil, "code":200, "message": "添加用户成功"}'
          schema:
            type: string
      tags:
      - 系统用户相关
    put:
      consumes:
      - application/json
      description: 更新用户
      parameters:
      - in: query
        name: name
        required: true
        type: string
      - in: query
        name: passone
        required: true
        type: string
      - in: query
        name: passtow
        required: true
        type: string
      - in: query
        name: role
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: '{"data": nil, "code":200, "message": "更新用户成功"}'
          schema:
            type: string
      tags:
      - 系统用户相关
swagger: "2.0"